+++ finagle-partitioning/src/main/scala/com/twitter/finagle/partitioning/FailureAccrualException.scala
@@ -1,0 +1,5 @@
+package com.twitter.finagle.partitioning
+
+import com.twitter.finagle._
+
+class FailureAccrualException(message: String) extends RequestException(message, cause = null)