+++ finagle-partitioning/src/main/scala/com/twitter/finagle/partitioning/PartitioningService.scala
@@ -1,4 +1,4 @@
-package com.twitter.finagle.memcached.partitioning
+package com.twitter.finagle.partitioning
 
 import com.twitter.finagle._
 import com.twitter.util._